Fixes in accounting
Expanded the slot to u32 to accommodate:
1. locally created DBs - fake slots taking up high 24bits
2. dependency DBs - true slots taking up low 8bits
3. stacks + globals (all nonDB accesses) - 0x0fffffff

+++ ocr/inc/ocr-edt-profiling.h
@@ -31,7 +31,7 @@ struct _profileStruct {
 
 struct _dbWeightStruct {
     u8     *fname;
-    u16    slots[2];    // Most dominant slots
+    u32    slots[2];    // Most dominant slots
     u8     weights[2];  // Corresponding weights to most dominant slots, %
     u8     sd[2];       // Corresponding sd to most dominant slots, 0-100
     u8     coverage;    // Coverage of most dominant 2 slots
